LayoutTests/imported/w3c/web-platform-tests/fetch/content-length/api-and-duplicate-headers.any-expected.txt


FAIL fetch() and duplicate Content-Length/Content-Type headers assert_equals: expected "text/plain, text/plain" but got "text/plain"
FAIL XMLHttpRequest and duplicate Content-Length/Content-Type headers assert_equals: expected "text/plain, text/plain" but got "text/plain"
